fix: PK 자료형 불일치 문제 해결

+++ tingting/src/main/java/com/alsif/tingting/concert/dto/concerthall/ConcertSeatBookRequestDto.java
@@ -21,5 +21,5 @@ import lombok.ToString;
 public class ConcertSeatBookRequestDto {
 
 	@Schema(description = "선택 좌석 PK 목록", type = "List<Integer>", example = "[154002,154003,154009]")
-	private List<Integer> seatSeqs;
+	private List<Long> seatSeqs;
 }
